Move helpers into support file, out of UI file
